requirements.txt
elasticsearch
psycopg2-binary
flask
certifi
